Move API functions to separate module
